pyproject.toml
[project]
name = "mcp-server-jupyter"
version = "0.1.0"
description = "MCP server for Jupyter notebooks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "mcp[cli]>=1.2.0",
    "nbclient>=0.10.2",
    "nbformat>=5.10.4",
    "pillow>=11.1.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_jupyter"]
